# spec/openapi.yaml
openapi: 3.1.0
info:
  title: x402 Mesh Inference Gateway
  version: 0.1.0
  summary: Pay-per-prompt OpenAI-compatible inference, settled in USDC on Algorand via x402.
  description: |
    A decentralized compute marketplace. Agents call an OpenAI-compatible
    `POST /v1/chat/completions` endpoint; the gateway answers `402 Payment Required` with an
    x402 v2 challenge, the caller re-sends the request carrying an `X-PAYMENT` header, and the
    gateway routes the prompt to the best available community GPU node.

    ## Payment flow

    1. `POST /v1/chat/completions` **without** `X-PAYMENT` → `402` with a `PaymentRequired`
       body. Its `accepts[]` carries the `exact` scheme on Algorand, the USDC ASA id, and the
       price in **atomic units** (`"2000"` = $0.0020 USDC, 6 decimals).
    2. The client builds an atomic Algorand transaction group (the facilitator's fee-payer
       transaction plus the client's ASA transfer), base64-encodes the resulting
       `PaymentPayload` as JSON, and sends it in the `X-PAYMENT` request header.
    3. The gateway verifies and settles through the facilitator, routes the prompt to a node,
       and returns the completion together with a `PAYMENT-RESPONSE` header carrying the
       base64-encoded `SettleResponse`.

    Payment is **gasless for the client**: the facilitator sponsors the ALGO transaction fee
    (advertised as `accepts[].extra.feePayer`), so the caller only signs the USDC transfer leg.

    ## Money representation

...

...

...
servers:
...
security: []
paths:
...
  /v1/nodes/register:
    post:
      operationId: registerNode
      tags: [nodes]
      summary: Register a compute node with the mesh.
      description: |
        Registration is self-signed: the operator signs the canonical serialization of the
        `registration` object with the Ed25519 key behind `operatorAddress`, and the gateway
        rejects any body whose signature, public key or address do not agree.

        The signing payload is
        `UTF-8("x402-mesh/node-registration/v1" + "\n" + canonicalJson(projection))`, where
        `projection` is the registration restricted to its eight fields (capabilities keep
        `model`, `contextWindow`, `pricePer1kTokensUsdc` and, when present, `quantization`),
        object keys are sorted, and `network` is normalized to canonical CAIP-2 first.

        `timestamp` must be within 120 seconds of the gateway clock and `nonce` must not have
        been seen before, which together bound registration replay.

        Registration also verifies **on chain** that `operatorAddress` has opted in to the USDC
        ASA. An account cannot receive an ASA it has not opted into, so a node that skips the
        opt-in would silently fail every payout. When `MESH_REQUIRE_USDC_OPT_IN` is enabled a
        node without the opt-in is stored but is not routable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignedNodeRegistration"
      responses:
        "201":
          description: Node registered (or its existing record replaced).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: |
            The `nodeId` is already registered to a different `operatorAddress`, or the `nonce`
            has already been used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshError"
              example:
                error:
                  code: validation_error
                  message: nodeId is already registered to another operator
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/nodes/{nodeId}/heartbeat:
    post:
      operationId: heartbeatNode
      tags: [nodes]
      summary: Report node liveness and load.
      description: |
        Sent by the node daemon every `MESH_HEARTBEAT_INTERVAL_MS` (15s by default). Carries the
        node's self-reported load so the router can shed traffic before a node saturates.

        Authenticated with the same Ed25519 envelope as registration, over the domain
        `x402-mesh/node-heartbeat/v1`. Missing heartbeats mark the node unhealthy and remove it
        from routing until it recovers.
      parameters:
        - $ref: "#/components/parameters/NodeIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignedNodeHeartbeat"
      responses:
        "200":
          description: Heartbeat accepted; the gateway's updated view of the node is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeHealth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: List registered compute nodes.
      description: |
        Public read model of the mesh. Useful for a leaderboard, for operators checking their
        own node's health, and for clients deciding whether a model has capacity before paying.
      parameters:
        - name: model
          in: query
          required: false
          description: Only return nodes advertising this exact model identifier.
          schema:
            type: string
            maxLength: 200
          example: llama3.1:8b
        - name: healthy
          in: query
          required: false
          description: When true, only nodes currently eligible for routing are returned.
          schema:
            type: boolean
            default: false
        - name: network
          in: query
          required: false
          description: Filter by settlement network, in either CAIP-2 encoding.
          schema:
            $ref: "#/components/schemas/Network"
        - name: limit
          in: query
          required: false
          description: Maximum number of nodes to return.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: The current node set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    x402Payment:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: |
        Base64 of the JSON `PaymentPayload` defined by x402 v2. Obtain the requirements to sign
        against from the `402` challenge on the same route. The legacy header name
        `Payment-Signature` is also accepted on input.
  parameters:
...
    NodeIdPath:
      name: nodeId
      in: path
      required: true
      description: The node's stable, operator-chosen identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
      example: gpu-fra-01
...
  responses:
...
    BadRequest:
      description: The request failed schema or freshness validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: validation_error
              message: invalid chat completion request
              details:
                fieldErrors:
                  messages:
                    - messages must not be empty
    Unauthorized:
      description: Ed25519 signature verification failed or the key does not match the address.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: auth_error
              message: signature does not verify against operatorAddress
    NotFound:
      description: No such node.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: validation_error
              message: unknown nodeId
...
    InternalError:
      description: Unexpected server-side failure. The message is deliberately opaque.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: internal_error
              message: Internal server error
  schemas:
    MeshError:
      type: object
      description: |
        Uniform error envelope. `code` is a stable machine-readable contract; `details` is
        sanitized before serialization, so it never contains key material or a stack trace.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - config_error
                - pricing_error
                - validation_error
                - payment_error
                - settlement_error
                - no_capacity
                - upstream_error
                - auth_error
                - rate_limited
                - internal_error
            message:
              type: string
              maxLength: 1024
            details:
              type: object
              additionalProperties: true
    Network:
      type: string
      description: |
        CAIP-2 network identifier. Both Algorand encodings are accepted on input; responses
        always carry the canonical truncated form.
      minLength: 3
      maxLength: 128
      pattern: "^[a-z0-9-]+:[A-Za-z0-9+/=_-]+$"
      examples:
        - algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73k
        - algorand:SGO1GKSzyE7IEPItTxCByw9x8FmnrCDe
        - algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8=
    AlgorandAddress:
      type: string
      description: 58-character RFC 4648 base32 Algorand address.
      pattern: "^[A-Z2-7]{58}$"
      examples:
        - MESHA65YISXOIW6STKMGJ5AFBZK2UJQNHR3LQZPG2LHIRB4YBPFJPT5WGU
    AtomicAmount:
      type: string
      description: |
        Non-negative integer USDC amount in atomic units (6 decimals) as a decimal string.
        `"2000"` is $0.0020. Never a float.
      pattern: "^\\d+$"
      examples:
        - "2000"
        - "1700"
        - "300"
    UsdcAmount:
      type: string
      description: |
        Non-negative decimal USDC amount with at most 6 fractional digits. Used for
        human-facing quotes; settlement always uses `AtomicAmount`.
      pattern: "^\\d+(\\.\\d{1,6})?$"
      examples:
        - "0.0020"
        - "0.000180"
...
    NodeCapability:
      type: object
      description: One model a node advertises, with its context window and quoted price.
      required: [model, contextWindow, pricePer1kTokensUsdc]
      additionalProperties: false
      properties:
        model:
          type: string
          minLength: 1
          maxLength: 200
          examples: [llama3.1:8b]
        contextWindow:
          type: integer
          description: Maximum prompt plus completion tokens accepted for this model.
          minimum: 1
          maximum: 10000000
          examples: [131072]
        pricePer1kTokensUsdc:
          $ref: "#/components/schemas/UsdcAmount"
        quantization:
          type: string
          minLength: 1
          maxLength: 64
          examples: [q4_K_M]
    NodeRegistration:
      type: object
      description: |
        A node's self-declaration. This exact object — projected onto these eight fields and
        canonically serialized — is what the operator signs.
      required:
        [nodeId, operatorAddress, endpoint, capabilities, network, version, timestamp, nonce]
      additionalProperties: false
      properties:
        nodeId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
          examples: [gpu-fra-01]
        operatorAddress:
          $ref: "#/components/schemas/AlgorandAddress"
        endpoint:
          type: string
          format: uri
          description: Absolute http(s) URL the gateway forwards inference requests to.
          maxLength: 512
          pattern: "^https?://"
          examples: [https://node-fra-01.example.com]
        capabilities:
          type: array
          minItems: 1
          maxItems: 64
          items:
            $ref: "#/components/schemas/NodeCapability"
        network:
          $ref: "#/components/schemas/Network"
        version:
          type: string
          minLength: 1
          maxLength: 32
          pattern: "^[0-9A-Za-z.+-]+$"
          examples: [0.1.0]
        timestamp:
          type: integer
          description: |
            Unix epoch milliseconds at signing time. Must be within 120000 ms of the gateway
            clock.
          minimum: 1
          examples: [1785772800000]
        nonce:
          type: string
          description: Single-use random hex, at least 128 bits.
          pattern: "^[0-9a-fA-F]{16,128}$"
          examples: [9f2c1d4e8a7b6053c1e2f4a6b8d0c3e5]
    SignedNodeRegistration:
      type: object
      description: A registration plus the Ed25519 proof that the operator key produced it.
      required: [registration, signature, publicKey]
      additionalProperties: false
      properties:
        registration:
          $ref: "#/components/schemas/NodeRegistration"
        signature:
          type: string
          description: Base64 Ed25519 signature (64 bytes) over the canonical signing bytes.
          contentEncoding: base64
          maxLength: 128
        publicKey:
          type: string
          description: |
            Base64 Ed25519 public key (32 bytes). Must hash to `registration.operatorAddress`.
          contentEncoding: base64
          maxLength: 64
    NodeHeartbeat:
      type: object
      description: |
        Self-reported liveness and load. Signed exactly like a registration but under the
        `x402-mesh/node-heartbeat/v1` domain.
      required: [nodeId, inFlight, latencyMsP50, latencyMsP95, timestamp, nonce]
      additionalProperties: false
      properties:
        nodeId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
        inFlight:
          type: integer
          description: Requests the node is currently serving.
          minimum: 0
        latencyMsP50:
          type: number
          minimum: 0
        latencyMsP95:
          type: number
          minimum: 0
        maxConcurrency:
          type: integer
          description: Updated concurrency ceiling, when the operator has changed it.
          minimum: 1
          maximum: 4096
        timestamp:
          type: integer
          minimum: 1
        nonce:
          type: string
          pattern: "^[0-9a-fA-F]{16,128}$"
    SignedNodeHeartbeat:
      type: object
      required: [heartbeat, signature, publicKey]
      additionalProperties: false
      properties:
        heartbeat:
          $ref: "#/components/schemas/NodeHeartbeat"
        signature:
          type: string
          contentEncoding: base64
          maxLength: 128
        publicKey:
          type: string
          contentEncoding: base64
          maxLength: 64
    NodeHealth:
      type: object
      description: Rolling liveness and quality signal the registry maintains per node.
      required:
        - nodeId
        - healthy
        - latencyMsP50
        - latencyMsP95
        - inFlight
        - maxConcurrency
        - uptimeRatio
        - qualityScore
        - consecutiveFailures
        - lastSeenAt
      additionalProperties: false
      properties:
        nodeId:
          type: string
          minLength: 1
          maxLength: 128
        healthy:
          type: boolean
          description: False once the node trips the failure threshold or misses heartbeats.
        latencyMsP50:
          type: number
          minimum: 0
        latencyMsP95:
          type: number
          minimum: 0
        inFlight:
          type: integer
          minimum: 0
        maxConcurrency:
          type: integer
          minimum: 1
          maximum: 4096
        uptimeRatio:
          type: number
          description: Fraction of heartbeat windows the node was reachable.
          minimum: 0
          maximum: 1
        qualityScore:
          type: number
          description: Composite routing score; higher is better.
          minimum: 0
          maximum: 1
        consecutiveFailures:
          type: integer
          minimum: 0
        lastSeenAt:
          type: integer
          description: Unix epoch milliseconds of the last successful heartbeat.
          minimum: 0
    NodeRecord:
      type: object
      description: Everything the registry stores about a node.
      required: [registration, health, usdcOptedIn, totalRequests, totalPaidAtomic, registeredAt]
      additionalProperties: false
      properties:
        registration:
          $ref: "#/components/schemas/NodeRegistration"
        health:
          $ref: "#/components/schemas/NodeHealth"
        usdcOptedIn:
          type: boolean
          description: |
            Verified on chain: the operator address holds the USDC ASA. A node without the
            opt-in cannot receive a payout, so it is not routable while
            `MESH_REQUIRE_USDC_OPT_IN` is enabled.
        totalRequests:
          type: integer
          minimum: 0
        totalPaidAtomic:
          $ref: "#/components/schemas/AtomicAmount"
        registeredAt:
          type: integer
          minimum: 0
    NodeSelection:
      type: object
      description: Router output — the chosen node plus why it won.
      required: [node, score, reason]
      additionalProperties: false
      properties:
        node:
          $ref: "#/components/schemas/NodeRecord"
        score:
          type: number
        reason:
          type: string
          maxLength: 512
          examples: [highest quality score among 3 nodes serving llama3.1:8b]
    NodeListResponse:
      type: object
      required: [nodes, count]
      additionalProperties: false
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeRecord"
        count:
          type: integer
          description: Number of nodes in `nodes`.
          minimum: 0
        selection:
          description: |
            Present only when the `model` query parameter was supplied and at least one node
            can serve it: the node the router would pick right now, and why. Purely
            informational — routing is re-evaluated at request time.
          $ref: "#/components/schemas/NodeSelection"
...